refactor: move Blog components to screens folder
