+++ src/hooks/useTwitterScript.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/ban-ts-comment */
 import { useEffect } from 'react';
 
 /* This loads all widgets from twitter if exists.
