+++ src/screens/Blog/components/FeaturedImage/FeaturedImage.tsx
@@ -3,10 +3,10 @@ import Img from 'gatsby-image';
 import { defineMessages } from 'react-intl';
 
 import { styled } from 'styles/styled';
+import { useBlogContext } from '@screens/Blog/hooks/useBlogContext';
 import { Frontmatter } from 'types';
 import { useIntl } from 'context/react-intl';
 import { Container } from 'components/Ui';
-import { useBlogContext } from '../blogContext';
 
 const StyledImg = styled(Img)`
   max-height: 600px;
