+++ src/screens/Blog/components/FeaturedImage/__mocks__/mockData.ts
