+++ src/screens/Blog/components/FeaturedImage/index.ts
