+++ src/screens/Blog/components/Header/Header.test.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 import { render } from 'test-utils';
 
 import { Header } from '.';
-import { BlogContextProvider } from '../blogContext';
-import { BlogPostContext } from '../types';
+import { BlogContextProvider } from '@screens/Blog/context';
+import { BlogPostContext } from '@screens/Blog/types';
 
 function renderWithContext(contextValues?: Partial<BlogPostContext>) {
   return render(
