+++ src/screens/Blog/components/MdxComponents/BigQuote.ts
