+++ src/screens/Blog/components/SeriesSection/__mocks__/mockedData.ts
